kopplat inloggning och skapa ny användare till api, ändrat interface idn till strängar
