+++ src/main.ts
@@ -1,6 +1,4 @@
-import p5 from 'p5';
 import { startPage } from './startpage';
-import { gameSketch } from './game';
 import './styles/main.scss';
 
 startPage();
